Fix more recovered bodies that diverged from retail
A second pass over the value, address, branch and return mismatches.
Behavior fixes:

- ResetSkillContribution stores to byte 1 of the skill record, not the
  active flag (documented as a retail bug).
- The party-selection controller raises the text buffer's geometry-dirty
  byte (+0x40), not the alternate-renderer byte (+0x41).
- The inspector load returns with _chdir(".."), not ".".
- FinishVideoPresentation restores DDSCL_NORMAL with a null window.
- The profession font's palette tables glow yellow, not blue.
- CreateMessageBox's image-name buffer is 32 bytes, so the retail-bug entry
  claiming an overrun was wrong and is removed.
- PromptRecordModeEntry and PlayCombatSound buffers are 100 bytes.

Type and signedness fixes:

- g_automap_page is unsigned; DamageCharacter's damage is unsigned.
- SetFlaggedRegionBounds takes ints.
- The NPC text-box callback passes zero-extended coordinates.
- CanCharacterUseItemEntry returns int.
- The targeting wrappers return their callee's byte unnormalized.
- CreateSoundEvent allocates strlen + 2.

The NPC notice callback's retail-bug entry is re-verified: only the left
release falls through, not the double click.

+++ src/wiz8/local_code/Controls.cpp
@@ -1051,11 +1051,10 @@ void W8TextControl::Invalidate(unsigned char immediate)
 }
 
 // FUNCTION: WIZ8 0x004f4600
-void W8TextControl::SetFlaggedRegionBounds(short left, short top, unsigned short right)
+void W8TextControl::SetFlaggedRegionBounds(int left, int top, int right)
 {
     if (m_region != -1 && m_pPanel != 0 && RegionHasFlags(m_region, 2)) {
-        SetRegionBounds(m_region, (unsigned short)((short)m_pPanel->origin_x + left),
-                        (unsigned short)((short)m_pPanel->origin_y + top), right, 0);
+        SetRegionBounds(m_region, m_pPanel->origin_x + left, m_pPanel->origin_y + top, right, 0);
     }
 }
